fix diorama control playing when game stopped

+++ map/Assets/Scripts/Components/Tile/TileData.cs
@@ -1,10 +1,9 @@
-#nullable enable
 public class TileData
 {
     public string? id;
     public int q;
     public int r;
     public int s;
-    public int height;
+    public float height;
     public string? color; // probably not right
 }
